+++ src/Splits.sol
@@ -16,9 +16,8 @@ struct SplitsReceiver {
 /// It's up to the caller to guarantee that this limit is never exceeded,
 /// failing to do so may result in a total protocol collapse.
 abstract contract Splits {
-    /// @notice Maximum number of splits receivers of a single user.
-    /// Limits cost of collecting.
-    uint32 internal constant _MAX_SPLITS_RECEIVERS = 200;
+    /// @notice Maximum number of splits receivers of a single user. Limits the cost of splitting.
+    uint256 internal constant _MAX_SPLITS_RECEIVERS = 200;
     /// @notice The total splits weight of a user
     uint32 internal constant _TOTAL_SPLITS_WEIGHT = 1_000_000;
     /// @notice The total amount the contract can keep track of each asset.
